+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/item/block/AbstractStorageContainerBlockItem.java
@@ -1,16 +1,21 @@
 package com.refinedmods.refinedstorage2.platform.api.item.block;
 
+import com.refinedmods.refinedstorage2.platform.api.PlatformApi;
 import com.refinedmods.refinedstorage2.platform.api.item.StorageContainerHelper;
+import com.refinedmods.refinedstorage2.platform.api.storage.StorageRepository;
 
+import java.util.List;
 import java.util.UUID;
 import javax.annotation.Nullable;
 
 import net.minecraft.core.BlockPos;
+import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -59,6 +64,23 @@ public abstract class AbstractStorageContainerBlockItem extends BlockItem {
                                       final ItemStack stack) {
         helper.getId(stack).ifPresent(id -> updateBlockEntityWithStorageId(pos, level.getBlockEntity(pos), id));
     }
+
+    @Override
+    public void appendHoverText(final ItemStack stack,
+                                @Nullable final Level level,
+                                final List<Component> tooltip,
+                                final TooltipFlag context) {
+        super.appendHoverText(stack, level, tooltip, context);
+        if (level == null) {
+            return;
+        }
+        final StorageRepository storageRepository = PlatformApi.INSTANCE.getStorageRepository(level);
+        helper.appendToTooltip(stack, storageRepository, tooltip, context, this::formatAmount, hasCapacity());
+    }
+
+    protected abstract boolean hasCapacity();
+
+    protected abstract String formatAmount(long amount);
 
     protected abstract ItemStack createPrimaryDisassemblyByproduct(int count);