+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/AbstractPlatform.java
@@ -45,7 +45,7 @@ public abstract class AbstractPlatform implements Platform {
     }
 
     @Override
-    public BucketAmountFormatting getBucketQuantityFormatter() {
+    public BucketAmountFormatting getBucketAmountFormatter() {
         return bucketAmountFormatting;
     }