+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/Platform.java
@@ -41,7 +41,7 @@ public interface Platform {
 
     TagKey<Item> getWrenchTag();
 
-    BucketAmountFormatting getBucketQuantityFormatter();
+    BucketAmountFormatting getBucketAmountFormatter();
 
     Config getConfig();