+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/storage/diskdrive/EmptyStorageDiskInfoAccessor.java
@@ -11,9 +11,4 @@ public class EmptyStorageDiskInfoAccessor implements StorageDiskInfoAccessor {
     public Optional<StorageInfo> getInfo(final ItemStack stack) {
         return Optional.empty();
     }
-
-    @Override
-    public boolean hasStacking(final ItemStack stack) {
-        return false;
-    }
 }