+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/resource/filter/fluid/FluidFilteredResource.java
@@ -39,7 +39,7 @@ public record FluidFilteredResource(FluidResource value, long amount) implements
 
     @Override
     public String getDisplayedAmount() {
-        return Platform.INSTANCE.getBucketQuantityFormatter().formatWithUnits(amount);
+        return Platform.INSTANCE.getBucketAmountFormatter().formatWithUnits(amount);
     }
 
     @Override