+++ CHANGELOG.md
@@ -7,6 +7,14 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+-  The upgrade slots now show their supported upgrades.
+
+### Removed
+
+-   Removed amount of stacks and max stacks stored on item storage tooltips.
+
 ## [2.0.0-milestone.2.7] - 2023-01-31
 
 ### Added